Refactor book module
